% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{AUC}
\alias{AUC}
\title{Calcula el area bajo la curva ROC 
 
Calcula una aproximacion del area bajo la curva ROC utilizando diferentes puntos de corte.}
\usage{
AUC(x, clase)
}
\arguments{
\item{x}{Vector numerico con los valores de la variable predictora.}

\item{clase}{Vector logico con las clases reales.}
}
\value{
Una aproximacion del area bajo la curva ROC.
}
\description{
Calcula el area bajo la curva ROC 
 
Calcula una aproximacion del area bajo la curva ROC utilizando diferentes puntos de corte.
}
