% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{calcular_metricas}
\alias{calcular_metricas}
\title{Calcula diferentes metricas para las variables de un conjunto de datos 
 
Calcula la varianza y el AUC para variables numericas y la entropia para variables categoricas.}
\usage{
calcular_metricas(dataset, clase)
}
\arguments{
\item{dataset}{Data frame que contiene las variables que se quieren analizar.}

\item{clase}{Nombre de la variable que contiene la clase.}
}
\value{
Un data frame con las metricas calculadas para cada variable.
}
\description{
Calcula diferentes metricas para las variables de un conjunto de datos 
 
Calcula la varianza y el AUC para variables numericas y la entropia para variables categoricas.
}
