% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{calcular_punto_ROC}
\alias{calcular_punto_ROC}
\title{Calcula un punto de la curva ROC}
\usage{
calcular_punto_ROC(datos, num.corte)
}
\arguments{
\item{datos}{Data frame que contiene las columnas valor y etiqueta.}

\item{num.corte}{Punto de corte utilizado para realizar la clasificacion.}
}
\value{
Un vector con los valores FPR y TPR.
}
\description{
Calcula la tasa de verdaderos positivos y la tasa de falsos positivos para un determinado punto de corte.
}
