% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{cor_dataset}
\alias{cor_dataset}
\title{Calcula la correlacion o informacion mutua entre todas las parejas de variables
 
Calcula una matriz aplicando cor_par() a cada pareja de variables.}
\usage{
cor_dataset(datos)
}
\arguments{
\item{datos}{Data frame cuyas variables se quieren comparar.}
}
\value{
Una matriz con las correlaciones o informaciones mutuas.
}
\description{
Calcula la correlacion o informacion mutua entre todas las parejas de variables
 
Calcula una matriz aplicando cor_par() a cada pareja de variables.
}
