% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{cor_par}
\alias{cor_par}
\title{Calcula la correlacion o la informacion mutua entre dos variables 
 
Calcula la correlacion de Pearson si ambas variables son numericas y la informacion mutua si ambas son factores.}
\usage{
cor_par(x, y)
}
\arguments{
\item{x}{Primera variable.}

\item{y}{Segunda variable.}
}
\value{
La correlacion de Pearson, la informacion mutua o NA si los tipos son mixtos.
}
\description{
Calcula la correlacion o la informacion mutua entre dos variables 
 
Calcula la correlacion de Pearson si ambas variables son numericas y la informacion mutua si ambas son factores.
}
