% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{discretize.dataset}
\alias{discretize.dataset}
\title{Discretiza las columnas numericas de un conjunto de datos}
\usage{
discretize.dataset(datos, cut.points)
}
\arguments{
\item{datos}{Data frame que se quiere discretizar.}

\item{cut.points}{Vector numerico con los puntos de corte.}
}
\value{
Un data frame con las columnas numericas discretizadas.
}
\description{
Aplica la funcion discretize() a todas las columnas numericas del conjunto de datos.
}
