% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{discretize}
\alias{discretize}
\title{Discretiza un vector numerico mediante puntos de corte}
\usage{
discretize(x, cut.points)
}
\arguments{
\item{x}{Vector numerico que se quiere discretizar.}

\item{cut.points}{Vector numerico con los puntos de corte.}
}
\value{
Un factor que indica el intervalo al que pertenece cada valor.
}
\description{
Divide un vector numerico en intervalos definidos por los puntos de corte.
}
