% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{discretizeEF.dataset}
\alias{discretizeEF.dataset}
\title{Discretiza las columnas numericas mediante intervalos de igual frecuencia 
 
Aplica discretizeEF() a todas las columnas numericas de un conjunto de datos.}
\usage{
discretizeEF.dataset(datos, num.bins = 3)
}
\arguments{
\item{datos}{Data frame que se quiere discretizar.}

\item{num.bins}{Numero de intervalos que se quieren crear.}
}
\value{
Un data frame con las columnas numericas discretizadas.
}
\description{
Discretiza las columnas numericas mediante intervalos de igual frecuencia 
 
Aplica discretizeEF() a todas las columnas numericas de un conjunto de datos.
}
