% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{discretizeEF}
\alias{discretizeEF}
\title{Discretiza un vector mediante intervalos de igual frecuencia}
\usage{
discretizeEF(x, num.bins = 3)
}
\arguments{
\item{x}{Vector numerico que se quiere discretizar.}

\item{num.bins}{Numero de intervalos que se quieren crear.}
}
\value{
Una lista con los valores categorizados y los grupos formados.
}
\description{
Ordena los valores y los divide en grupos con un numero aproximadamente igual de observaciones.
}
