% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{discretizeEW.dataset}
\alias{discretizeEW.dataset}
\title{Discretiza las columnas numericas mediante intervalos de igual amplitud 
 
Aplica discretizeEW() a todas las columnas numericas de un conjunto de datos.}
\usage{
discretizeEW.dataset(datos, num.bins)
}
\arguments{
\item{datos}{Data frame que se quiere discretizar.}

\item{num.bins}{Numero de intervalos que se quieren crear.}
}
\value{
Un data frame con las columnas numericas discretizadas.
}
\description{
Discretiza las columnas numericas mediante intervalos de igual amplitud 
 
Aplica discretizeEW() a todas las columnas numericas de un conjunto de datos.
}
