% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{discretizeEW}
\alias{discretizeEW}
\title{Discretiza un vector mediante intervalos de igual amplitud 
 
Divide el rango de un vector numerico en un numero determinado de intervalos con la misma amplitud.}
\usage{
discretizeEW(x, num.bins)
}
\arguments{
\item{x}{Vector numerico que se quiere discretizar.}

\item{num.bins}{Numero de intervalos que se quieren crear.}
}
\value{
Una lista con los valores categorizados, los intervalos y los puntos de corte.
}
\description{
Discretiza un vector mediante intervalos de igual amplitud 
 
Divide el rango de un vector numerico en un numero determinado de intervalos con la misma amplitud.
}
