% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{entropy}
\alias{entropy}
\title{Calcula la entropia de una variable categorica}
\usage{
entropy(x)
}
\arguments{
\item{x}{Variable de tipo factor.}
}
\value{
El valor de la entropia.
}
\description{
Calcula la entropia de una variable de tipo factor a partir de sus frecuencias.
}
