% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{estandarizar.dataset}
\alias{estandarizar.dataset}
\title{Estandariza las columnas numericas de un conjunto de datos 
 
Aplica la funcion estandarizar() a todas las columnas numericas.}
\usage{
estandarizar.dataset(datos)
}
\arguments{
\item{datos}{Data frame que se quiere estandarizar.}
}
\value{
Un data frame con las columnas numericas estandarizadas.
}
\description{
Estandariza las columnas numericas de un conjunto de datos 
 
Aplica la funcion estandarizar() a todas las columnas numericas.
}
