% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{estandarizar}
\alias{estandarizar}
\title{Estandariza un vector numerico}
\usage{
estandarizar(x)
}
\arguments{
\item{x}{Vector numerico que se quiere estandarizar.}
}
\value{
El vector estandarizado.
}
\description{
Estandariza un vector utilizando su media y desviacion estandar.
}
