% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{filtrado_variables}
\alias{filtrado_variables}
\title{Filtra variables segun diferentes metricas}
\usage{
filtrado_variables(data, umbral_varianza, umbral_auc, umbral_entropia)
}
\arguments{
\item{data}{Data frame que contiene las metricas calculadas.}

\item{umbral_varianza}{Umbral minimo de varianza para las variables continuas.}

\item{umbral_auc}{Umbral minimo de AUC para las variables continuas.}

\item{umbral_entropia}{Umbral minimo de entropia para las variables discretas.}
}
\value{
Un vector con los nombres de las variables seleccionadas.
}
\description{
Selecciona las variables que superan los umbrales establecidos para la varianza, el AUC y la entropia.
}
