% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{grafica_ROC}
\alias{grafica_ROC}
\title{Representa graficamente una curva ROC}
\usage{
grafica_ROC(x, clase)
}
\arguments{
\item{x}{Vector numerico con los valores de la variable predictora.}

\item{clase}{Vector logico con las clases reales.}
}
\value{
No devuelve ningun objeto; genera una representacion grafica de la curva ROC.
}
\description{
Calcula diferentes puntos de una curva ROC y los representa graficamente.
}
