% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{integraOptimizada}
\alias{integraOptimizada}
\title{Calcula una integral mediante la regla del trapecio 
 
Realiza una aproximacion numerica de una integral utilizando la regla del trapecio.}
\usage{
integraOptimizada(x, y)
}
\arguments{
\item{x}{Vector con los valores del eje x.}

\item{y}{Vector con los valores del eje y.}
}
\value{
Una aproximacion numerica de la integral.
}
\description{
Calcula una integral mediante la regla del trapecio 
 
Realiza una aproximacion numerica de una integral utilizando la regla del trapecio.
}
