% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{normalizar.dataset}
\alias{normalizar.dataset}
\title{Normaliza las columnas numericas de un conjunto de datos 
 
Aplica la funcion normalizar() a todas las columnas numericas.}
\usage{
normalizar.dataset(datos)
}
\arguments{
\item{datos}{Data frame que se quiere normalizar.}
}
\value{
Un data frame con las columnas numericas normalizadas.
}
\description{
Normaliza las columnas numericas de un conjunto de datos 
 
Aplica la funcion normalizar() a todas las columnas numericas.
}
