% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{normalizar}
\alias{normalizar}
\title{Normaliza un vector numerico 
 
Transforma los valores de un vector numerico al intervalo entre 0 y 1.}
\usage{
normalizar(x)
}
\arguments{
\item{x}{Vector numerico que se quiere normalizar.}
}
\value{
El vector normalizado.
}
\description{
Normaliza un vector numerico 
 
Transforma los valores de un vector numerico al intervalo entre 0 y 1.
}
