% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{plot_correlaciones}
\alias{plot_correlaciones}
\title{Representa graficamente una matriz de correlaciones}
\usage{
plot_correlaciones(datos)
}
\arguments{
\item{datos}{Data frame cuyas variables se quieren comparar.}
}
\value{
Un grafico de tipo mapa de calor.
}
\description{
Calcula la matriz de correlaciones o informacion mutua y la representa mediante un mapa de calor.
}
