% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Funciones.R
\name{varianza}
\alias{varianza}
\title{Calcula la varianza de un vector numerico}
\usage{
varianza(x)
}
\arguments{
\item{x}{Vector numerico}
}
\value{
El valor de la varianza.
}
\description{
Calcula la varianza muestral de un vector numerico
}
